Move Intermediate representation tests into their own module
